requirements.txt
freeplane_io >=0.10.0
pylatex >=1.4.2
peek-python >=1.7.0
cairosvg >=2.7.1